Here separate classes for ex1, ex2 and ex3 are removed and combined into one for cucumber,steps and run classes
