+++ src/test/java/com/epam/tc/hw5/TestRun.java
@@ -1,8 +1,8 @@
-package com.epam.tc.hw5.ex2;
+package com.epam.tc.hw5;
 
 import io.cucumber.testng.AbstractTestNGCucumberTests;
 import io.cucumber.testng.CucumberOptions;
 
 @CucumberOptions
-public class ExerciseTwoTestRun extends AbstractTestNGCucumberTests {
+public class TestRun extends AbstractTestNGCucumberTests {
 }
