+++ src/test/java/com/epam/tc/hw5/cucumber/hooks/CucumberHook.java
@@ -1,6 +1,6 @@
-package com.epam.tc.hw5.ex2.cucumber.hooks;
+package com.epam.tc.hw5.cucumber.hooks;
 
-import com.epam.tc.hw5.ex1.cucumber.context.TestContext;
+import com.epam.tc.hw5.cucumber.context.TestContext;
 import io.cucumber.java.After;
 import io.cucumber.java.Before;
 import io.github.bonigarcia.wdm.WebDriverManager;
