--- src/test/java/com/epam/tc/hw5/ex1/ExerciseOneTestRun.java
@@ -1,8 +1,0 @@
-package com.epam.tc.hw5.ex1;
-
-import io.cucumber.testng.AbstractTestNGCucumberTests;
-import io.cucumber.testng.CucumberOptions;
-
-@CucumberOptions
-public class ExerciseOneTestRun extends AbstractTestNGCucumberTests {
-}
