--- src/test/java/com/epam/tc/hw5/ex3/ExerciseThreeTest.java
@@ -1,8 +1,0 @@
-package com.epam.tc.hw5.ex3;
-
-import io.cucumber.testng.AbstractTestNGCucumberTests;
-import io.cucumber.testng.CucumberOptions;
-
-@CucumberOptions
-public class ExerciseThreeTest extends AbstractTestNGCucumberTests {
-}
