+++ src/test/java/com/epam/tc/hw5/steps/AbstractStep.java
@@ -1,15 +1,19 @@
-package com.epam.tc.hw5.ex1.steps;
+package com.epam.tc.hw5.steps;
 
-import com.epam.tc.hw5.ex1.cucumber.context.TestContext;
+import com.epam.tc.hw5.cucumber.context.TestContext;
 import com.epam.tc.hw5.pages.DifferentPage;
 import com.epam.tc.hw5.pages.GdiGitHubSite;
+import com.epam.tc.hw5.pages.UserTablePage;
+import com.epam.tc.hw5.pages.components.HeaderComponents;
 import com.epam.tc.hw5.pages.components.LeftSideMenuComponents;
 import org.assertj.core.api.SoftAssertions;
 import org.openqa.selenium.WebDriver;
 
 public abstract class AbstractStep {
 
-    protected static GdiGitHubSite epamPage;
+    protected static GdiGitHubSite gdiGitHubSite;
+    protected static HeaderComponents headerComponents;
+    protected static UserTablePage userTablePage;
     protected static LeftSideMenuComponents leftSideMenuComponents;
     protected static DifferentPage differentPage;
 
@@ -17,7 +21,8 @@ public abstract class AbstractStep {
         WebDriver webDriver = TestContext.getInstance().getTestObject("web_driver");
         SoftAssertions softAssertions = TestContext.getInstance().getTestObject("soft_assertion");
 
-        epamPage = new GdiGitHubSite(webDriver, softAssertions);
+        gdiGitHubSite = new GdiGitHubSite(webDriver, softAssertions);
+        headerComponents = new HeaderComponents(webDriver, softAssertions);
         leftSideMenuComponents = new LeftSideMenuComponents(webDriver, softAssertions);
     }
 }
