+++ src/test/java/com/epam/tc/hw5/steps/AssertionStep.java
@@ -1,10 +1,30 @@
-package com.epam.tc.hw5.ex2.steps;
+package com.epam.tc.hw5.steps;
 
 import java.util.List;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.Then;
 
 public class AssertionStep extends AbstractStep {
+
+    @Then("for checkbox 'Water' there is an individual log row and value is corresponded to the status of checkbox")
+    public void verifyCheckboxSelectedWater() {
+        differentPage.verifySelectedCheckBoxWater();
+    }
+
+    @Then("for checkbox 'Wind' there is an individual log row and value is corresponded to the status of checkbox")
+    public void verifySelectedCheckboxWin() {
+        differentPage.verifySelectedCheckBoxWind();
+    }
+
+    @Then("for radio button there is a log row and value is corresponded to the status of radio button")
+    public void verifyRadioSelected() {
+        differentPage.verifySelectedRadioSelen();
+    }
+
+    @Then("for dropdown there is a log row and value is corresponded to the selected value")
+    public void verifyDropdownSelected() {
+        differentPage.verifySelectedDropDownYellow();
+    }
 
     @Then("{string} page should be opened")
     public void verifyUserTablePageOpen(String title) {
@@ -41,5 +61,10 @@ public class AssertionStep extends AbstractStep {
     public void verifyDroplistColumnValues(DataTable dropdownValues) {
         List<String> informationForUserRoman = dropdownValues.asList();
         userTablePage.verifyDroplistInformationForUserRoman(informationForUserRoman);
+    }
+
+    @Then("I log row has \"Vip: condition changed to true\" text in log section")
+    public void verifyVipLog() {
+        userTablePage.verifyLogForCheckboxIsDisplayed();
     }
 }
