+++ src/test/resources/com/epam/tc/hw5/ex1/SelectFunctionTest.feature
@@ -1,8 +1,7 @@
 Feature: Test on Different Page function for select elements
-
-  Scenario: Select checkbox, radio and element on dropdown list
-    Given I open Home Page
-    And I perform login
+  Scenario: Select checkbox, radio and element on dropdown list test
+    Given I open JDI GitHub site
+    And I login as user 'Roman Iovlev'
     And I open through the header menu Service -> Different Elements Page
     When I select checkbox 'Water'
     And I select checkbox 'Wind'
