+++ src/test/resources/com/epam/tc/hw5/ex3/SelectFunctionInUserTablePageTest.feature
@@ -1,7 +1,7 @@
 Feature: Test GDI GitHub site function to select 'vip' in User Table Page
-  Scenario: Select function in User Table Page Test
+  Scenario: Function to select 'vip' in User Table Page Test
     Given I open JDI GitHub site
-    And I login as user "Roman Iovlev"
+    And I login as user 'Roman Iovlev'
     And I click on "Service" button in Header
     And I click on "User Table" button in Service dropdown
     When I select 'vip' checkbox for 'Sergey Ivan'
