+++ src/test/resources/testing/HW5TestRun.xml
@@ -5,9 +5,7 @@
   </listeners>
   <test name="RunEx1andEx2Tests">
     <classes>
-      <class name="com.epam.tc.hw5.ex1.ExerciseOneTestRun"/>
-      <class name="com.epam.tc.hw5.ex2.ExerciseTwoTestRun"/>
-      <class name="com.epam.tc.hw5.ex3.ExerciseThreeTest"/>
+      <class name="com.epam.tc.hw5.TestRun"/>
     </classes>
   </test>
 </suite>
